# api/openapi-full.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/payment-groups": {
      "post": {
        "operationId": "PaymentGroupsController_create",
        "summary": "Create payment group",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreatePaymentGroupDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The payment group has been successfully created."
          },
          "400": {
            "description": "Invalid input, unable to create the payment group."
          }
        },
        "tags": ["Payment Groups"],
        "security": [{ "bearer": [] }]
      },
      "get": {
        "operationId": "PaymentGroupsController_findAll",
        "summary": "List payment groups",
        "parameters": [],
        "responses": {
          "200": { "description": "Successfully retrieved the payment groups." }
        },
        "tags": ["Payment Groups"],
        "security": [{ "bearer": [] }]
      }
    },
    "/payment-groups/{id}": {
      "get": {
        "operationId": "PaymentGroupsController_findOne",
        "summary": "Retrieve payment group",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Successfully retrieved the payment group." },
          "404": {
            "description": "No payment group found for the provided id."
          }
        },
        "tags": ["Payment Groups"],
        "security": [{ "bearer": [] }]
      },
      "patch": {
        "operationId": "PaymentGroupsController_update",
        "summary": "Update payment group",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdatePaymentGroupDto" }
            }
          }
        },
        "responses": {
          "200": { "description": "Successfully updated the payment group." },
          "400": {
            "description": "Invalid input, unable to update the payment group."
          }
        },
        "tags": ["Payment Groups"],
        "security": [{ "bearer": [] }]
      }
    },
    "/payment-groups/{id}/participants": {
      "get": {
        "operationId": "PaymentGroupsController_participants",
        "summary": "Retrieve participants",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Successfully retrieved the participants." },
          "404": {
            "description": "No participants found for the provided payment group id."
          }
        },
        "tags": ["Payment Groups"],
        "security": [{ "bearer": [] }]
      }
    },
    "/payment-groups/{id}/cancel": {
      "post": {
        "operationId": "PaymentGroupsController_cancel",
        "summary": "Cancel payment group",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Successfully canceled the payment group." },
          "400": {
            "description": "Invalid request, unable to cancel the payment group."
          }
        },
        "tags": ["Payment Groups"],
        "security": [{ "bearer": [] }]
      }
    },
    "/payment-groups/{id}/expire": {
      "post": {
        "operationId": "PaymentGroupsController_expire",
        "summary": "Expire payment group",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Successfully expired the payment group." },
          "400": {
            "description": "Invalid request, unable to expire the payment group."
          }
        },
        "tags": ["Payment Groups"],
        "security": [{ "bearer": [] }]
      }
    },
    "/payment-groups/{id}/validate": {
      "post": {
        "operationId": "PaymentGroupsController_validate",
        "summary": "Validate payment group",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidatePaymentGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Successfully validated the payment group." },
          "400": {
            "description": "Invalid input, unable to validate the payment group."
          }
        },
        "tags": ["Payment Groups"],
        "security": [{ "bearer": [] }]
      }
    },
...
  },
  "info": {
    "title": "kohortPay",
...
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    { "url": "https://api.kohortpay.dev", "description": "staging" },
    { "url": "http://localhost:3000/", "description": "Development" },
    { "url": "https://api.kohortpay.com", "description": "Production" }
  ],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" },
      "basic": { "type": "http", "scheme": "basic" }
    },
    "schemas": {
...
      "CreatePaymentGroupDto": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string",
            "description": "The ID of the customer.",
            "example": "cus_IzkjlvAhdjzjht3"
          },
          "paymentIntentId": {
            "type": "string",
            "description": "The ID of the payment intent.",
            "example": "pi_1JYLo8KerLxWZaQtys6ZQ1xR"
          },
          "metadata": {
            "type": "object",
            "description": "Additional metadata for the payment group.",
            "example": { "order_id": "ord_1JYLo8KerLxWZaQtys6ZQ1xS" }
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string",
            "description": "The expiration date and time of the payment group.",
            "example": "2023-11-30T14:24:01.171Z"
          }
        },
        "required": ["customerId", "paymentIntentId"]
      },
      "UpdatePaymentGroupDto": {
        "type": "object",
        "properties": {
          "metadata": {
            "type": "object",
            "description": "Additional metadata for the payment group update.",
            "example": { "order_id": "ord_1JYLo8KerLxWZaQtys6ZQ1xT" }
          }
        },
        "required": ["metadata"]
      },
      "ValidatePaymentGroupDto": {
        "type": "object",
        "properties": {
          "customerEmail": {
            "type": "string",
            "description": "The email address of the customer.",
            "example": "customer@gmail.com"
          }
        },
        "required": ["customerEmail"]
      },
...
    }
  }
}